+++ CHANGELOG.md
@@ -65,3 +65,4 @@
 # v2.2.0
 
 - Added plugin configuration for support of special chars (like umlauts) in the Source URL
+- Fixed error on creating/editing a redirect with an empty Source-/Target-URL